SOSI: fix 2 MSVC warnings about loss of precision in implicit cast, and constify char* parameter, use CPLAtof()
git-svn-id: https://svn.osgeo.org/gdal/trunk@34419 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/sosi/ogrsosilayer.cpp
@@ -329,7 +329,7 @@ OGRFeature *OGRSOSILayer::GetNextFeature() {
                     int date[6];
                     SOSITypeToDateTime(tokens[k], date);
                     if (date[0]>0) 
-                      poFeature->SetField( iHNr, date[0], date[1], date[2], date[3], date[4], date[5], 1);
+                      poFeature->SetField( iHNr, date[0], date[1], date[2], date[3], date[4], static_cast<float>(date[5]), 1);
                     break;
                   }
                   case OFTReal: {
